Feat/user crud

+++ src/main/java/com/vhskillpro/backend/modules/role/RoleService.java
@@ -6,26 +6,31 @@ import com.vhskillpro.backend.modules.role.dto.RoleCreateDTO;
 import com.vhskillpro.backend.modules.role.dto.RoleDTO;
 import com.vhskillpro.backend.modules.role.dto.RoleDetailDTO;
 import com.vhskillpro.backend.modules.role.dto.RoleUpdateDTO;
+import com.vhskillpro.backend.modules.user.UserRepository;
 import jakarta.transaction.Transactional;
 import java.util.Optional;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
 @Service
 public class RoleService {
   private ModelMapper modelMapper;
   private RoleRepository roleRepository;
+  private UserRepository userRepository;
   private PermissionRepository permissionRepository;
 
   public RoleService(
       ModelMapper modelMapper,
       RoleRepository roleRepository,
+      UserRepository userRepository,
       PermissionRepository permissionRepository) {
     this.roleRepository = roleRepository;
     this.modelMapper = modelMapper;
+    this.userRepository = userRepository;
     this.permissionRepository = permissionRepository;
   }
 
@@ -38,6 +43,7 @@ public class RoleService {
    * @param pageable the pagination information
    * @return a page of RoleDTO objects matching the search criteria
    */
+  @PreAuthorize("hasAnyAuthority('all:all', 'role:read')")
   @Transactional
   public Page<RoleDTO> findAll(String keyword, Pageable pageable) {
     return roleRepository
@@ -52,6 +58,7 @@ public class RoleService {
    * @return an {@link Optional} containing the mapped {@link RoleDetailDTO} if found, or an empty
    *     Optional if not found
    */
+  @PreAuthorize("hasAnyAuthority('all:all', 'role:read')")
   @Transactional
   public Optional<RoleDetailDTO> findById(Long id) {
     return roleRepository.findById(id).map(role -> modelMapper.map(role, RoleDetailDTO.class));
@@ -64,6 +71,7 @@ public class RoleService {
    * @param roleCreateDTO the data transfer object containing role details and permission IDs
    * @return the detailed data transfer object of the saved role
    */
+  @PreAuthorize("hasAnyAuthority('all:all', 'role:create')")
   @Transactional
   public RoleDetailDTO create(RoleCreateDTO roleCreateDTO) {
     Role role =
@@ -89,6 +97,7 @@ public class RoleService {
    * @return the updated role details as a {@link RoleDetailDTO}
    * @throws AppException if the role with the specified ID is not found
    */
+  @PreAuthorize("hasAnyAuthority('all:all', 'role:update')")
   @Transactional
   public RoleDetailDTO update(Long id, RoleUpdateDTO roleUpdateDTO) {
     Role role =
@@ -117,12 +126,18 @@ public class RoleService {
    * @return {@code true} if the role was successfully deleted
    * @throws AppException if the role with the specified ID does not exist
    */
+  @PreAuthorize("hasAnyAuthority('all:all', 'role:delete')")
   @Transactional
-  public boolean delete(Long roleId) {
+  public void delete(Long roleId) {
     if (!roleRepository.existsById(roleId)) {
       throw new AppException(HttpStatus.NOT_FOUND, RoleMessages.ROLE_NOT_FOUND.getMessage());
     }
+
+    // Check if the role is associated with any users
+    if (userRepository.existsByRoleId(roleId)) {
+      throw new AppException(HttpStatus.CONFLICT, RoleMessages.ROLE_DELETE_CONFLICT.getMessage());
+    }
+
     roleRepository.deleteById(roleId);
-    return true;
   }
 }
